% Generated by roxygen2 (4.0.2): do not edit by hand
\name{czy_poziomy_sa_okej}
\alias{czy_poziomy_sa_okej}
\title{Sprawdzanie czy poziomy spełniają określone kryteria}
\usage{
czy_poziomy_sa_okej(tablicaPoziomow, maxLPozWyk = 5, minLiczebnPozWyk = 100,
  minOdsPozWyk = 0.05)
}
\arguments{
\item{tablicaPoziomow}{tablica poziomów}

\item{maxLPozWyk}{maksymalna liczba poziomów pytania}

\item{minLiczebnPozWyk}{minimalna liczebność obserwacji w każdym z poziomów}

\item{minOdsPozWyk}{minimalny odsetek obserwacji, które ma zawierać każdy poziom.}
}
\value{
wartość logiczna
}
\description{
Funkcja na podstawie tablicy poziomów zwróconych przez funkcję \code{\link[base]{table}} określa,
czy spełnione są określone warunki.
}

