% Generated by roxygen2 (4.0.1): do not edit by hand
\name{detekcja_wypracowan}
\alias{detekcja_wypracowan}
\title{Detekcja wypracowań}
\usage{
detekcja_wypracowan(dane)
}
\arguments{
\item{dane}{ramka danych z wynikami.}
}
\value{
Lista.
}
\description{
Funkcja zwraca listę, której pierwszy element 'wspolne' zawiera indeksy kolumn,
które nie zawierają braków danych. Z kolei drugi element 'wypracowania' zawiera listę indeksów kolumn,
dla których braki danych występują dla tych samych obserwacji (wierszów).
}

