\name{EWDskalowanie-package}
\alias{EWDskalowanie-package}
\alias{EWDskalowanie}
\docType{package}
\title{
Pakiet zawiera funkcje automatyzujące proces skalowania IRT z użyciem progamu Mplus.
}
\description{
Używanie pakietu koncentruje się wokół funkcji code{\link{skaluj}}.
W przyszłości, o ile będzie czas i chęci, planowane jest sporotowanie funkcji code{\link{skaluj}} również na pakiet \code{mirt}.
}
\details{
\tabular{ll}{
Package: \tab EWDskalowanie\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-03-27\cr
License: \tab LGPL-3\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Tomasz Żółtak <tomek@zozlak.org>

Maintainer: Tomasz Żółtak <tomek@zozlak.org>
}
\keyword{ package }
