% Generated by roxygen2 (4.0.1): do not edit by hand
\name{generuj_poziom_pytania}
\alias{generuj_poziom_pytania}
\title{Generowanie poziomów dla (pseudo)kryterium}
\usage{
generuj_poziom_pytania(vec, nazwa_pytania, zrodloDanychODBC = "EWD")
}
\arguments{
\item{vec}{wektor z odpowiedziami na (pseudo)kryterium}

\item{nazwa_pytania}{nazwa pytania}

\item{zrodloDanychODBC}{źródło danych ODBC}
}
\value{
wektor poziomów
}
\description{
Funkcja na podstawie bazy danych oraz samych danych uczniów określa liczbę potencjalnych poziomów dla podanego (pseudo)kryterium.
}

