% Generated by roxygen2 (4.0.1): do not edit by hand
\name{initializuj_skalowanie_polich}
\alias{initializuj_skalowanie_polich}
\title{Zainiciowanie obliczeń.}
\usage{
initializuj_skalowanie_polich(dane, procedura, korelacjaWiazki, nazwaSkalowania,
  minMiara)
}
\arguments{
\item{dane}{ramka danych zawierająca wyniki egzaminów.}

\item{procedura}{procedura skalowania egzaminów.}

\item{korelacjaWiazki}{tablica zawierająca kolumny kr1 oraz kr2 z wartościami kryteriów do połącznia
oraz kolumnę numer określającą kolejność obliczeń korelacji.}

\item{nazwaSkalowania}{ciąg znaków opisujący skalowanie.}

\item{minMiara}{patrz definicja \link{skaluj_polichorycznie}.}
}
\value{
Patrz opis funkcji \code{\link{skaluj_polichorycznie}}.
}
\description{
Funkacja tworzy obiekt klasy WynikSkalowania używany do dalszych obliczeń.
}

