% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kolejny_krok_polich}
\alias{kolejny_krok_polich}
\title{Jedno skalowanie z połączeniem danych}
\usage{
kolejny_krok_polich(wynikSkalowania)
}
\arguments{
\item{wynikSkalowania}{element klasy WynikSkalowania.}
}
\value{
Element klasy WynikSkalowania.
}
\description{
Funkcja wykonuje jednen krok skalowania wraz z pobocznymi operacjami, jak połączenie danych
oraz zwiększenie indeksu.
}
\details{
Patrz detale funkcji \code{\link{skaluj_polichorycznie}}.
}

