% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kryteria_do_polaczenia}
\alias{kryteria_do_polaczenia}
\title{Wskazywanie kryteriów do połączenia}
\usage{
kryteria_do_polaczenia(rodzajEgzaminu, czescEgzaminu, rokEgzaminu,
  zrodloDanychODBC = "EWD_grzes", czyPolski = grepl("polski",
  czescEgzaminu), czesciWypracowania = 6)
}
\arguments{
\item{rodzajEgzaminu}{rodzaj egzaminu}

\item{czescEgzaminu}{część egzaminu}

\item{rokEgzaminu}{rok egzaminu}

\item{zrodloDanychODBC}{żródło ODBC}

\item{czyPolski}{zmienna logiczna określająca, czy mamy do czynienia z językiem polskim}

\item{czesciWypracowania}{parametr liczbowy nie ignorowany, gdy zmienna czyPolski=TRUE.
Określa liczbę części wypracowania (domyślnie 6) i ignoruje w wyniku wszystkie kryteria, który opisy powtarzają się tyle razy co wskazuje ten parametr.}
}
\value{
Funkcja zwraca listę numerów kryteriów do połączenia.
}
\description{
Funkcja zwraca listę numerów kryteriów, które powinny być połaczone.
}
\examples{
\dontrun{
rodzajEgzaminu = "matura"
czescEgzaminu = "informatyka podstawowa"
rokEgzaminu = 2014
kryteria_do_polaczenia(rodzajEgzaminu, czescEgzaminu, rokEgzaminu)
}
}

