% Generated by roxygen2 (4.0.2): do not edit by hand
\name{kryteria_z_nazw}
\alias{kryteria_z_nazw}
\title{Kryteria z nazw}
\usage{
kryteria_z_nazw(nazwy)
}
\arguments{
\item{nazwy}{nazwy kryteriów.}
}
\value{
Funkcja zwraca wektor liczb oznaczających numery kryteriów.
}
\description{
Funkcja wyznacza numer kryteriów z ich nazw.
}

