% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lacz_kryteria_zadan}
\alias{lacz_kryteria_zadan}
\title{Laczenie kryteriow wedlug numerow zadan}
\usage{
lacz_kryteria_zadan(rodzajEgzaminu, czescEgzaminu, rokEgzaminu,
  zrodloDanychODBC = "EWD")
}
\arguments{
\item{rodzajEgzaminu}{ciąg znaków - rodzaj egzaminu}

\item{czescEgzaminu}{ciąg znaków - część egzaminu}

\item{rokEgzaminu}{liczba całkowita - rok egzaminu}

\item{zrodloDanychODBC}{żródło ODBC}
}
\value{
Funkcja zwraca data frame, który można uzyć jako argument \code{elementy}
funkcji \code{edytuj_skale()} z pakietu \code{ZPD}.
}
\description{
Funkcja łączy kryteria oceny poszczególnych zadań (wyróżnionych na podstawie numeracji
w arkuszu egzaminu) w pseudokryteria.
}
\examples{
\dontrun{
rodzajEgzaminu = "matura"
czescEgzaminu = "informatyka podstawowa"
rokEgzaminu = 2014
kryteria_do_polaczenia(rodzajEgzaminu, czescEgzaminu, rokEgzaminu)
}
}

