% Generated by roxygen2 (4.0.2): do not edit by hand
\name{laczenie_wykresy_zapis}
\alias{laczenie_wykresy_zapis}
\title{Łączenie oraz zapis wykresów do plików}
\usage{
laczenie_wykresy_zapis(dane, plikDoZapisu, folderWykresy, folderWyniki,
  czy2Konstr)
}
\arguments{
\item{dane}{ramka danych z wynikami części egzaminu.}

\item{plikDoZapisu}{nazwa pliku, która posłuży do zbudowania nazw plików z wynikami skalowania
oraz plików png wykresów.}

\item{folderWykresy}{folder zapisu wykresów.}

\item{folderWyniki}{folder zapisu wyników skalowania.}

\item{czy2Konstr}{zmienna boolowska. Jeżeli wynosi TRUE to łączenie będzie odbywać się na podstawie wartości drugiego konstruktu.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja wykonuje łączenie, skalowania oraz zapisuje wyniki do plików.
}

