% Generated by roxygen2 (4.0.2): do not edit by hand
\name{okresl_skrocenie}
\alias{okresl_skrocenie}
\title{Funkcja określająca skrótcenie}
\usage{
okresl_skrocenie(vec, poziomy, maxLPozWyk = 5, minLiczebnPozWyk = 100,
  minOdsPozWyk = 0.05)
}
\arguments{
\item{vec}{wektor z wynikami danego pytania}

\item{poziomy}{wektor z dopuszczalnymi poziomami dla danego pytania}

\item{maxLPozWyk}{maksymalna liczba poziomów pytania}

\item{minLiczebnPozWyk}{minimalna liczebność obserwacji w każdym z poziomów}

\item{minOdsPozWyk}{minimalny odsetek obserwacji, które ma zawierać każdy poziom.}
}
\value{
Funkcja zwraca skrócony wektor - factor, którego atrybut 'polaczenie' zawiera tabelę z zastępowanymi i zastępującym poziomami.
}
\description{
Pomocnicza funkcja, która łączy dwa poziomy sprawdzając wcześniej, czy spełniają one dopuszczalne warunki (patrz pis funkcji \code{\link{czy_poziomy_sa_okej}}).
}

