% Generated by roxygen2 (4.0.2): do not edit by hand
\name{okresl_wzor_skracania}
\alias{okresl_wzor_skracania}
\title{Okreslenie wzoru skrocenia skali oceny}
\usage{
okresl_wzor_skracania(x, mozliweWartosci, maxLPozWyk = 5,
  minLiczebnPozWyk = 100, minOdsPozWyk = 0.05)
}
\arguments{
\item{x}{wektor z wynikami danego pytania}

\item{mozliweWartosci}{wektor z dopuszczalnymi poziomami dla danego pytania}

\item{maxLPozWyk}{maksymalna liczba poziomów pytania}

\item{minLiczebnPozWyk}{minimalna liczebność obserwacji w każdym z poziomów}

\item{minOdsPozWyk}{minimalny odsetek obserwacji, które ma zawierać każdy poziom.}
}
\value{
Funkcja zwraca trzyelementową listę, której elementy zawierają:
\itemize{
  \item{\code{przedSkroceniem} wartość parametru \code{mozliweWartosci},}
  \item{\code{poSkroceniu} wartości po skróceniu, odpowiadające wartościom pierwszego
        elementu,}
  \item{\code{rozkladPrzed} rozkład \code{x},}
  \item{\code{rozkladPo} rozkład \code{x} po skróceniu skali.}
}
}
\description{
Funkcja dokonuje skrótu skali oceny (pseudo)kryterium w oparciu o rozkład jego
wyników w grupie kalibracyjnej i informację o wszystkich możliwych wartościach,
jakie może przyjąć dane (pseudo)kryterium.
}

