% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pobierz_mozliwe_wartosci}
\alias{pobierz_mozliwe_wartosci}
\title{Wartosci mozliwe do przyjecia przez (pseudo)kryterium}
\usage{
pobierz_mozliwe_wartosci(nazwyKryteriow, zrodloDanychODBC = "EWD")
}
\arguments{
\item{nazwyKryteriow}{wektor tekstowy z nazwami postaci "k_idKryterium" lub
"p_idPseudokryterium" (typowo nazwy zmiennych z data frame'a zwróconego przez
funkcję \code{obierz_czesc_egzaminu()} z pakietu \code{ZPD}).}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do
bazy (domyślnie "EWD")}
}
\value{
lista wektorów liczbowych
}
\description{
Funkcja pobiera z bazy i zwraca informacje o wszystkich dopuszczalnych wartościach
kryteriów i/lub pseudokryteriów.
}

