% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pobierz_parametry_skalowania}
\alias{pobierz_parametry_skalowania}
\title{Pobieranie parametrow skalowan.}
\usage{
pobierz_parametry_skalowania(nazwa_skali = NULL, id_testu = NULL,
  opis_skalowania = ".*", zrodloDanychODBC = "ewd_grzes",
  parametryzacja = "baza")
}
\arguments{
\item{nazwa_skali}{nazwa skali. Kiedy nazwa skali przymuje wartość NULL nie jest brana pod
uwagę w wyszukiwaniu. Wartość domyślna to NULL.}

\item{id_testu}{id testu.Kiedy id testu przymuje wartość NULL nie jest brana pod
uwagę w wyszukiwaniu. Wartość domyślna to NULL.}

\item{opis_skalowania}{wyrażenie regularne określające opis skalowania. Domyślna wartość to '\%'.}

\item{zrodloDanychODBC}{string określający źródło danych. docelowa wartość domyślna to 'EWD'. Obecnie 'ewd_grzes'.}

\item{parametryzacja}{parametr określający format zwracanego wyniku. Domyślna wartość to 'baza'.
Inna możliwa wartość to 'mplus'.}
}
\value{
W przyopadku użycia parametryzacji 'baza', funkcja zwraca listę taką, że:
\itemize{
   \item każdy element listy opisuje parametry innego skalowania;
   \item każdy element listy ma też przypisane jako atrybuty wartości kolumn: 'skalowanie', 'opis', 'estymacja';
   \item zwracana lista ma przypisane jako atrybuty wartości wszystkich kolumn z tablicy 'skale';
}
W przyopadku użycia parametryzacji 'mplus' funkcja zwraca parametry skalowania w formie ramki danych, która jest
w postaci zwracanej przez funkcję skaluj().
}
\description{
Funkcja pobiera parametry skalowań o podanej nazwie skali i numerze testu oraz których
opis skalowań spełnia podane wyrażenie regularne.
}
\examples{
\dontrun{
id_testu = 1128
nazwa_skali = "ktt;1128"
ret = pobierz_parametry_skalowania(nazwa_skali,id_testu)
}
}

