% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/korelacje_polichoryczne.R
\name{pobierz_wiazki_pytania_kryteria}
\alias{pobierz_wiazki_pytania_kryteria}
\title{Pobierz: wiązki, pytania, kryteria}
\usage{
pobierz_wiazki_pytania_kryteria(kryteria, zrodloODBC = "EWD")
}
\arguments{
\item{kryteria}{wektor zawietający numery kryteriów.}

\item{zrodloODBC}{źródło danych ODBC.}
}
\value{
Funkcja zwraca ramkę danych.
}
\description{
Funkcja pobiera dane dla zdefiniowanych kryteriów.
}

