% Generated by roxygen2 (4.0.1): do not edit by hand
\name{podmien_wartosci_lista}
\alias{podmien_wartosci_lista}
\title{Zmiana wartości jednego z węzłów listy}
\usage{
podmien_wartosci_lista(wezel, nowaWartosc, doZmiany)
}
\arguments{
\item{wezel}{lista zawierająca, której węzły chcemy zmienić.}

\item{nowaWartosc}{nowa wartość dla zmienianego elementu}

\item{doZmiany}{nazwa węzła do zmiany.}
}
\value{
Zmodyfikowana lista.
}
\description{
Funkcja zmienia rekurencyjnie wartości węzłów o nazwie określonej w parametrze "doZmiany".
}

