% Generated by roxygen2 (4.0.2): do not edit by hand
\name{polacz_dane}
\alias{polacz_dane}
\title{Łączenie kryteriów}
\usage{
polacz_dane(dane, korelacjaWiazki = NULL, index = 0)
}
\arguments{
\item{dane}{ramka danych zawierająca wyniki egzaminów.}

\item{korelacjaWiazki}{tablica zawierająca kolumny kr1 oraz kr2 z wartościami kryteriów do połącznia
oraz kolumnę numer określającą kolejność obliczeń korelacji.}

\item{index}{numer określający wiersz tablicy korelacjaWiazki, który wskazuje kryteria do połączenia.}
}
\value{
Funkcja zwaca ramkę danych, na której wykonano operację połączenia.
}
\description{
Funkcja łączy w danych dwie kolumny zawierające wyniki kryteriów.
}

