% Generated by roxygen2 (4.0.2): do not edit by hand
\name{polacz_kryteria_mirt}
\alias{polacz_kryteria_mirt}
\title{Łączenie kryteriów przy wykorzystniu funkcji mirt.}
\usage{
polacz_kryteria_mirt(dane, mirtSummary, wiazki_pyt_kryt = NULL, prog = NULL,
  h2 = FALSE)
}
\arguments{
\item{dane}{dane w formie tabeli do obliczeń skalowania.}

\item{mirtSummary}{rezultat wywołania summary dla obiektu mirt.}

\item{wiazki_pyt_kryt}{tablica zawierająca informacje o wiązkach, pytaniach, kryteriach.
Parametr może być efektem działania funkcji \code{\link{pobierz_wiazki_pytania_kryteria}}.}

\item{prog}{liczba określająca próg, powyżej którego ma dochodzić do połączeń.}

\item{h2}{zmienna logiczna, która określa czy wykorzystywać miarę h2
czy wartość drugiego konstruktu do łączeń kryteriów.}
}
\value{
Funkcja zwraca listę zawierającą:
\itemize{
\item \code{dane} - wynik połączenia;
\item \code{polaczenie} - informacja o wykonanym połączeniu;
\item \code{wyniki} - tablica zawirejąca statystyki kryteriów wykorzystywane przy wyborze kryteriów do połączenia;
}
}
\description{
Funkcja łączy kryteria korzystając z wyników skalowania funkcją mirt
}
\details{
Jeżeli parametr prog wynosi NULL to łączone są dwa kryteria o najwyższej mierze określanej przez parametr h2.
Jeżeli parametr wiazki_pyt_kryt równaja się NULL to podział na wiązki nie jest wykorzystywany przy łączeniu.
}

