% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skalowanieWieluCzesci.R
\name{polacz_nieliczne}
\alias{polacz_nieliczne}
\title{Łączenie poziomów odpowiedzi}
\usage{
polacz_nieliczne(vec, poziomy = 5)
}
\arguments{
\item{vec}{ramka danych z wynikami części egzaminu.}

\item{poziomy}{liczba określająca, ile poziomów ma być w wynikowym wektorze.}
}
\value{
Funkcja zwraca factor o liczbie poziomów o jeden miejszą niż parametr 'vec'.
}
\description{
Funkcja łączy najmniej liczny poziom z mniej licznym z poziomów sąsiednich.
Funkcja wykonuje łączenie do momentu osiągnięcia zadanej liczby poziomów.
}

