% Generated by roxygen2 (4.0.1): do not edit by hand
\name{polacz_poziom}
\alias{polacz_poziom}
\title{Łączenie dwóch poziomów dla wektora klasy factor}
\usage{
polacz_poziom(vecFactor, maxLPozWyk = 5, minLiczebnPozWyk = 100,
  minOdsPozWyk = 0.05)
}
\arguments{
\item{vecFactor}{wektor factorów}

\item{maxLPozWyk}{maksymalna liczba poziomów pytania}

\item{minLiczebnPozWyk}{minimalna liczebność obserwacji w każdym z poziomów}

\item{minOdsPozWyk}{minimalny odsetek obserwacji, które ma zawierać każdy poziom.}
}
\value{
Funkcja zwraca wektor klasy factor, który dodatkowo posiada atrybuty. Pierwszy 'czy_poziomy_sa_okej' określa, czy poziomy spełniają zadane warunki.
Drugi atrybut 'polaczenie' zawiera tablicę z połączeniami. Jeżeli wektor wejściowy zawierał ten argument to funkcja dopisuje do niego wiersz z wykonanym połączniem
oraz aktualizuje poprzenie wiersze tego atrybutu.
}
\description{
Pomocnicza funkcja, która łączy dwa poziomy sprawdzając wcześniej, czy spełniają one dopuszczalne warunki (patrz opis funkcji \code{\link{czy_poziomy_sa_okej}}).
}

