% Generated by roxygen2 (4.0.1): do not edit by hand
\name{policz_korelacje_wiazki}
\alias{policz_korelacje_wiazki}
\title{Liczenie korelacji polichorycznych}
\usage{
policz_korelacje_wiazki(dane_pytan, wiazki_pyt_kryt)
}
\arguments{
\item{dane_pytan}{ramka danych zawierająca wyniki kryteriów.}

\item{wiazki_pyt_kryt}{ramka danych, która zawiera informacje o przynależności pytań do poszczególnych wiązek.}
}
\value{
Funkcja zwraca ramkę danych, w której znajdują się obliczone korelacje polichoryczne.
}
\description{
Funkcja liczy korelacje polichoryczne dla kryteriów zebranych w wiązki.
}

