% Generated by roxygen2 (4.0.1): do not edit by hand
\name{procedura_eg_hum}
\alias{procedura_eg_hum}
\title{Procedury skalowania egzaminow.}
\usage{
procedura_eg_hum(nazwyZmiennych, parametryGH = NULL, processors = 3)
}
\arguments{
\item{nazwyZmiennych}{nazwy zmiennych z data.frame'a z danymi, na których ma być prowadzona estymacja}

\item{parametryGH}{wartości startowe parametrów.}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista, która zostanie użyta jako argument \code{opisProcedury} funkcji \code{\link{skaluj}}
}
\description{
Procedura skalowania części humanistycznej egzaminu gimnazjalnego.
Funkcja przygotowuj opis procedury skalowania do użycia przez funkcję \code{\link{skaluj}}.
}

