% Generated by roxygen2 (4.0.1): do not edit by hand
\name{procedura_eg_sf_matura}
\alias{procedura_eg_sf_matura}
\title{Procedury skalowania egzaminow.}
\usage{
procedura_eg_sf_matura(nazwyZmiennych, parametryGH = NULL,
  parametryGMP = NULL, processors = 3)
}
\arguments{
\item{nazwyZmiennych}{nazwy zmiennych z data.frame'a z danymi, na których ma być prowadzona estymacja}

\item{parametryGH}{data.frame z parametrami zadań z części hum. egzaminu gimnazjalnego}

\item{parametryGMP}{data.frame z parametrami zadań z części mat.-przyr. egzaminu gimnazjalnego}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista, która zostanie użyta jako argument \code{opisProcedury} funkcji \code{\link{skaluj}}
}
\description{
Procedura skalowania egzaminu gimnazjalnego - w formule do 2011 r. na potrzeby danych do wskaźników matrualnych.
Funkcja przygotowuj opis procedury skalowania do użycia przez funkcję \code{\link{skaluj}}.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}
}

