% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/procedury_matura_wielogrupowe.R
\name{procedura_matura_hm_gr_wszyscy}
\alias{procedura_matura_hm_gr_wszyscy}
\title{Procedury skalowania egzaminow.}
\usage{
procedura_matura_hm_gr_wszyscy(nazwyZmiennych, parametryJpol, parametryHum,
  processors = 3)
}
\arguments{
\item{nazwyZmiennych}{nazwy zmiennych z data.frame'a z danymi, na których ma być prowadzona estymacja}

\item{parametryJpol}{data.frame z parametrami zadań z przedmiotów humanistycznych}

\item{parametryHum}{data.frame z parametrami zadań z języka polskiego}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista, która zostanie użyta jako argument \code{opisProcedury} funkcji \code{\link{skaluj}}
}
\description{
Procedura skalowania matury - przedmioty humanistyczne z uwzględnieniem laureatów.
Funkcja przygotowuj opis procedury skalowania do użycia przez funkcję \code{\link{skaluj}}.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}
}

