% Generated by roxygen2 (4.0.2): do not edit by hand
\name{procedura_matura_m_rasch_wszyscy}
\alias{procedura_matura_m_rasch_wszyscy}
\title{Procedury skalowania egzaminow.}
\usage{
procedura_matura_m_rasch_wszyscy(nazwyZmiennych, parametryMat, processors = 3)
}
\arguments{
\item{nazwyZmiennych}{nazwy zmiennych z data.frame'a z danymi, na których ma być prowadzona estymacja}

\item{parametryMat}{data.frame z parametrami zadań z matematyki}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista, która zostanie użyta jako argument \code{opisProcedury} funkcji \code{\link{skaluj}}
}
\description{
Procedura skalowania matury - matematyka modelem Rascha na potrzeby Kalkulatora z uwzględnieniem laureatów.
Funkcja przygotowuj opis procedury skalowania do użycia przez funkcję \code{\link{skaluj}}.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}
}

