% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedury_matura_wielogrupowe.R
\name{procedura_matura_mp_gr_wszyscy}
\alias{procedura_matura_mp_gr_wszyscy}
\title{Procedury skalowania egzaminow.}
\usage{
procedura_matura_mp_gr_wszyscy(nazwyZmiennych, parametryMat, parametryMatPrzyr,
  processors = 3)
}
\arguments{
\item{nazwyZmiennych}{nazwy zmiennych z data.frame'a z danymi, na których ma być prowadzona estymacja}

\item{parametryMat}{data.frame z parametrami zadań z matematyki}

\item{parametryMatPrzyr}{data.frame z parametrami zadań z przedmiotów matematyczno-przyrodniczych}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista, która zostanie użyta jako argument \code{opisProcedury} funkcji \code{\link{skaluj}}
}
\description{
Procedura skalowania matury - przedmioty matematyczno-przyrodnicze z uwzględnieniem laureatów.
Funkcja przygotowuj opis procedury skalowania do użycia przez funkcję \code{\link{skaluj}}.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}
}

