% Generated by roxygen2 (4.0.2): do not edit by hand
\name{przygotuj_kryteria}
\alias{przygotuj_kryteria}
\title{Przygotowanie obiektu do edycji skali}
\usage{
przygotuj_kryteria(kryt, zrodloDanychODBC = "EWD")
}
\arguments{
\item{kryt}{lista zwrócona przez funkcję \code{\link{kryteria_do_polaczenia}}}

\item{zrodloDanychODBC}{żródło danych}
}
\value{
Funkcja zwraca ramkę danych, której wiersze albo zawierają id istniejącego pseudokryterium albo id wszystkich kryteriów, które mają zostać połączone w jedno pseudokryterium.
}
\description{
Funkcja przygotowuje ramkę danych, która może zostać wygodnie użyta do edycji skali przez funkcję \code{\link[ZPD]{edytuj_skale}} z pakietu ZPD.
}

