% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skalowanieWieluCzesci.R
\name{skalowanieWieluCzesci}
\alias{skalowanieWieluCzesci}
\title{Funkcja skalujaca i łącząca wiele części egzaminów}
\usage{
skalowanieWieluCzesci(plik, folderDane, folderWyniki, folderWykresy,
  odswiez = FALSE, czy2Konstr = TRUE)
}
\arguments{
\item{plik}{nazwa pliku csv zawierający tabelę z wynikami części egzaminu.}

\item{folderDane}{ścieżka do folderu, który zawiera plik, którego nazwa jest wskazana parametrem 'plik'.}

\item{folderWyniki}{ścieżka do folderu, gdzie zostanie zapisany wynik skalowania i łączenia.}

\item{folderWykresy}{ścieżka do folderu, gdzie zostaną zapisane wykresy wyników skalowania.}

\item{odswiez}{zmienna boolowska określająca, czy należy powtórzyć obliczenia, jeżeli w katalogu znajduje się plik o nazwie wyniku algorytmu.}

\item{czy2Konstr}{zmienna boolowska. Jeżeli wynosi TRUE to łączenie będzie odbywać się na podstawie wartości drugiego konstruktu.
W przeciwnym wypadku wybór kryteriów do łączenia będzie odbywał się na podstawie wartości korelacji polichorycznych.}
}
\value{
Funkcja nic nie zwraca.
}
\description{
Funkcja zapisuje do wskazanego folderu wyniki skalowania
}
\details{
Funkcja do folderu folderWyniki zapisuje rezultat działania funkcji skaluj_laczenie_mirt lub skaluj_polichorycznie.
Dodatkowo do folderWykresy zapisuje też histogramy i wykresy gęstości.
Nazwa pliku z wynikami skalowań i łączeń powstaje przez dodanie '.rkp' (korelacje polichoryczne) lub '.r2k' (dwa konstrukty)
do parametru plik. W sytuacjach egzaminu, który składa się z wielu części przez rozszerzeniem pliku doklejane są nazwy konstruktów,
które dla tego skalowania nie przyjmója wartości NA. Ma to miejsce na przykład dla WOS'u oraz j.polskiegod.
}

