% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skaluj_egz_gimn_rasch.R
\name{skaluj_egz_gimn_rasch}
\alias{skaluj_egz_gimn_rasch}
\title{Procedury skalowania egzaminow.}
\usage{
skaluj_egz_gimn_rasch(daneWzorcowe, daneWszyscy, processors = 2)
}
\arguments{
\item{daneWzorcowe}{lista data frame'ów zawierających dane do skalowania wzorcowego}

\item{daneWszyscy}{lista data frame'ów zawierających zawierający dane wszystkich
zdających (do wyliczenia oszacowań umiejętności na podstawe parametrów ze skalowania
wzorcowego)}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista z elementami:
\itemize{
  \item{\code{usunieteKryteria} lista wektorów tekstowych z nazwami (pseudo)kryteriów,
   które zostały usunięte podczas skalowania wzorcowego.}
  \item{\code{parametry} lista data frame'ów z wyestymowanymi parametrami modeli w ich
        ostatecznej postaci (tj. takiej, jak w ostatnim kroku skalowania wzorcowego
        i w jedynym kroku skalowania na wszystkich zdających).}
  \item{\code{oszacowania} lista data frame'ów zawierających id_obserwacji i wyliczone
        oszacowania umiejętności dla wszystkich zdających.}
  \item{\code{rzetelnoscEmpiryczna} rzetelność wyliczona na podstawie oszacowań ze
        skalowania wzorcowego (jako wariancja oszacowań EAP).}
  \item{\code{mapowanie} data frame zawierający wzorzec przekodowania
        wyników surowych na wyniki wyskalowane.}
  \item{\code{odsUtraconejWariancji} odsetek wariancji przewidywania mierzonej cechy
        utracony w wyniku uśrednienia oszacowań w funkcji sumy punktów.}
}
}
\description{
Funkcja przeprowadza skalowanie części mat.-przyr. egzaminu gimnazjalnego z użyciem
modelu Rascha (na potrzeby maturalnego Kalkulatora EWD).
}
\details{
Argumenty (poza ostatnim) muszą być listami (mogą być jednoelementowe), których
elementy mają nazwy ze zbioru: \code{"gm", "gm_p", "gm_m"} i są data frame'ami
zawierającymi dane z wynikami odpowiednich części egzaminu gimnazjalnego.

W przypadku nowej struktury egzaminu gimnazjalnego wystarczy podać wyniki testów,
dane do wyskalowania części jako całości zostaną połączone automatycznie, wewnątrz
funkcji.

Schemat przekodowania sum punktów na oszacowania umiejętności wyliczany jest na
podstawie danych wzorcowych, przy pomocy funkcji \code{\link{przewidywanie_rasch}},
a następnie na jego podstawie przypisywane są wartości przewidywane wszystkim zdającym.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}, \code{\link{procedura_1k_1w}}, \code{\link{przewidywanie_rasch}}
}

