% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skaluj_krok}
\alias{skaluj_krok}
\title{Łączenie danych kryteriów oraz ich skalowanie.}
\usage{
skaluj_krok(dane, proceduraEG, opisSkalowania, korelacjaWiazki, index,
  wartosciStartowe = NULL)
}
\arguments{
\item{dane}{ramka daych kryteriów (przed połączeniem).}

\item{proceduraEG}{procedura skalowania.}

\item{opisSkalowania}{ciąg znaków opisujący skalowanie.}

\item{korelacjaWiazki}{ramka danych zawierająca korelacje polichoryczne.}

\item{index}{numer wiersza ramki korelacjeWiazki, który zawiera informacje o połączeniach.}

\item{wartosciStartowe}{wartości startowe skalowania. Domyślna wartość to NULL.}
}
\value{
Funkcja zwraca listę z wynikami skalowania, wartości startowe, dane po połączeniu, procedurę oraz połączenie.
}
\description{
Funkcja, która wykonuje łączenie danych oraz skalowanie z użyciem Mplusa.
}

