% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skaluj_krok_mirt}
\alias{skaluj_krok_mirt}
\title{Łączenie danych kryteriów oraz ich skalowanie.}
\usage{
skaluj_krok_mirt(dane, opisSkalowania, korelacjaWiazki, index)
}
\arguments{
\item{dane}{ramka daych kryteriów (przed połączeniem).}

\item{opisSkalowania}{ciąg znaków opisujący skalowanie.}

\item{korelacjaWiazki}{ramka danych zawierająca korelacje polichoryczne.}

\item{index}{numer wiersza ramki korelacjeWiazki, który zawiera informacje o połączeniach.}
}
\value{
Funkcja zwraca listę z wynikami skalowania, wartości startowe, dane po połączeniu, procedurę oraz połączenie.
}
\description{
Funkcja, która wykonuje łączenie danych oraz skalowanie w wykorzystaniem pakietu mirt.
}

