% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mirt_laczenie.R
\name{skaluj_laczenie_mirt}
\alias{skaluj_laczenie_mirt}
\title{Łączenie i skalowanie na podstawie wyników funkcji mirt.}
\usage{
skaluj_laczenie_mirt(dane, wiazki_pyt_kryt = NULL, prog = NULL,
  h2 = FALSE, maxIter = ifelse(is.null(prog), 10, 100),
  czyLiczyc1 = FALSE)
}
\arguments{
\item{dane}{dane.}

\item{wiazki_pyt_kryt}{tablica zawierająca informacje o wiązkach, pytaniach, kryteriach.
Patrz \code{\link{pobierz_wiazki_pytania_kryteria}}.}

\item{prog}{liczba określająca próg, powyżej którego ma dochodzić do połączeń.}

\item{h2}{zmienna logiczna, która określa czy wykorzystywać miarę h2
czy wartość drugiego konstruktu do łączeń kryteriów.}

\item{maxIter}{maksymalna liczba iteracji.}

\item{czyLiczyc1}{zmienna logiczna określająca
czy w wynikach działania funkcji ma się znaleźć rezultat skalowania funkcją mirt dla jednego kostruktu
czy dla dwóch konstruktów. Ustawienie parametru na FALSE przyśpiesza obliczenia.}
}
\value{
Funkcja zwraca listę zawierającą:
\itemize{
\item \code{wyniki} - wyniki skalowań;
\item \code{polaczenia} - informacja o wykonanych polaczeniach;
\item \code{czasy} - informacja o czasach oszczegónych iteracji;
\item \code{dane} - dane użyte do wykonania ostatniego skalowania;
}
}
\description{
Funkcja skaluje i łączy kryteria korzystając ze skalowania funkcją mirt modelami IRT zawierającymi 2 zmienne ukryte.
}
\details{
Jeżeli parametr prog wynosi NULL to łączone są dwa kryteria o najwyższej mierze określanej przez parametr h2.
Jeżeli parametr wiazki_pyt_kryt równaja się NULL to podział na wiązki nie jest wykorzystywany przy łączeniu.
}

