% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skaluj_matura.R
\name{skaluj_matura}
\alias{skaluj_matura}
\title{Procedury skalowania egzaminow.}
\usage{
skaluj_matura(daneWzorcowe, daneWszyscy, czesciEgzaminu = NULL,
  sufiksy = NULL, processors = 2)
}
\arguments{
\item{daneWzorcowe}{lista data frame'ów zawierających dane do skalowania wzorcowego}

\item{daneWszyscy}{lista data frame'ów zawierających zawierający dane wszystkich
zdających (do wyliczenia oszacowań umiejętności na podstawe parametrów ze skalowania
wzorcowego)}

\item{czesciEgzaminu}{opcjonalnie lista wektorów tekstowych zawierających nazwy
elementów argumnetów \code{daneWzorcowe} i \code{daneWszyscy}, które mają być
uwzględnione przy wyliczaniu poszczególnych wskaźników wyników matury (każdy element
listy opisuje jeden wskaźnik); jeśli \code{NULL}, wyliczone zostaną cztery domyślne
wskaźniki}

\item{sufiksy}{opcjonalnie lista ciągów znaków definiująca sufiksy odpowiadające
poszczególnym częściom egzaminu (elementom argumentów \code{daneWzorcowe}
i \code{daneWszyscy}); nazwa elementu listy definiuje część egzaminu, wartość
elementu to sufiks}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista z elementami:
\itemize{
  \item{\code{usunieteKryteria} lista wektorów tekstowych z nazwami (pseudo)kryteriów,
   które zostały usunięte podczas skalowania wzorcowego.}
  \item{\code{parametry} lista data frame'ów z wyestymowanymi parametrami modeli w ich
        ostatecznej postaci (tj. takiej, jak w ostatnim kroku skalowania wzorcowego
        i w jedynym kroku skalowania na wszystkich zdających).}
  \item{\code{oszacowania} lista data frame'ów zawierających id_obserwacji i wyliczone
        oszacowania umiejętności dla wszystkich zdających.}
  \item{\code{grupy} data frame zawierający mapowanie wartości zmiennych definiujących
        podział na grupy na wartości technicznej zmiennej grupującej (\code{gr_tmp1}),
        wykorzystywanej w estymacji przez funkcję \code{\link{skaluj}}.}
}
}
\description{
Funkcja przeprowadza skalowanie matury na potrzeby trzyletnich wskaźników maturalnych.
Domyślnie wyliczne są cztery wskaźniki: 1) humanistyczny (j. polski, historia, WOS),
2) matematyczno-przyrodniczy (matematyka, biologia, chemia, fizyka, geografia,
informatyka), 3) polonistyczny (j. polski), 4) matematyczny (matematyka).
}
\details{
Argumenty (poza ostatnim) muszą być listami (mogą być jednoelementowe), których
elementy są data frame'ami zawierającymi dane z wynikami odpowiednich części matury.

\bold{Model skalowania:} jednowymiarowe, wielogrupowe modele 2PL/SGR, z grupami
zdefiniowanymi przez: 1) w przypadku modeli dla j. polskiego i matematyki: typ szkoły
(LO/T) i pisanie lub nie poziomu rozszerzonego lub 2) w przypadku modeli obejmujących
wiele przedmiotów: typ szkoły (LO/T).

\bold{Uwaga:} dopisywanie parametrów selekcji związanych z wyborem tematu wypracowania
(polski PP i PR, historia PR, WOS PR) jest naklepane ostrym brutem i jest raczej
nieodporne na zmiany w formule matury w zakresie możliwości wyboru tematu wypracowań.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}, \code{\link{procedura_1k_1w}},
\code{\link{sprawdz_wyniki_skalowania}}
}

