% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skaluj_matura_rasch}
\alias{skaluj_matura_rasch}
\title{Procedury skalowania egzaminow.}
\usage{
skaluj_matura_rasch(daneWzorcowe, daneWszyscy, processors = 2)
}
\arguments{
\item{daneWzorcowe}{lista data frame'ów zawierających dane do skalowania wzorcowego}

\item{daneWszyscy}{lista data frame'ów zawierających zawierający dane wszystkich
zdających (do wyliczenia oszacowań umiejętności na podstawe parametrów ze skalowania
wzorcowego)}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista, której każdy element opisuje wyniki dla innego typu szkół (spośród występujących
w danych) i składa się z elementów:
\itemize{
  \item{\code{usunieteKryteria} lista wektorów tekstowych z nazwami (pseudo)kryteriów,
   które zostały usunięte podczas skalowania wzorcowego.}
  \item{\code{parametry} lista data frame'ów z wyestymowanymi parametrami modeli w ich
        ostatecznej postaci (tj. takiej, jak w ostatnim kroku skalowania wzorcowego
        i w jedynym kroku skalowania na wszystkich zdających).}
  \item{\code{oszacowania} lista data frame'ów zawierających id_obserwacji i wyliczone
        oszacowania umiejętności dla wszystkich zdających. \bold{Wyniki zwracane są
        w formie wystandaryzowanej do (0, 1) na danych wzorocwych.}}
  \item{\code{mapowanie} data frame zawierający wzorzec przekodowania
        wyników surowych na wyniki wyskalowane.}
  \item{\code{odsUtraconejWariancji} odsetek wariancji przewidywania mierzonej cechy
        utracony w wyniku uśrednienia oszacowań w funkcji sumy punktów.}
  \item{\code{grupy} data frame zawierający mapowanie wartości zmiennych definiujących
        podział na grupy (\code{typ_szkoly} i \code{pr}) na wartości technicznej
        zmiennej grupującej (\code{gr_tmp1}), wykorzystywanej w estymacji przez funkcję
        \code{\link{skaluj}}.}
}
}
\description{
Funkcja przeprowadza skalowanie matury z matematyki łącznie na poziomach podstawowym
i rozszerzonym, z wykorzystaniem wielogrupowego modelu Rascha (na potrzeby maturalnego
Kalkulatora EWD). Grupy definiowane są przez wybór poziomu rozszerzonego i typ szkoły
(LO/T).
}
\details{
Argumenty (poza ostatnim) muszą być listami (mogą być jednoelementowe), których
elementy mają nazwy ze zbioru: \code{"matematyka podstawowa", "matematyka rozszerzona"}
i są data frame'ami zawierającymi dane z wynikami odpowiednich części matury.

Schemat przekodowania sum punktów na oszacowania umiejętności wyliczany jest na
podstawie danych wzorcowych, przy pomocy funkcji \code{\link{przewidywanie_rasch}},
a następnie na jego podstawie przypisywane są wartości przewidywane wszystkim zdającym.

Ponieważ różni zdający rozwiązywali różne zestawy zadań, a do tego model jest
wielogrupowy, funkcja nie wylicza (i nie zwraca) wartości rzetleności empirycznej.

\bold{Model skalowania:} jednowymiarowy, wielogrupowy model Rascha (Rasch/Rasch Graded
Response), z grupami zdefiniowanymi przez typ szkoły (LO/T) i pisanie lub nie poziomu
rozszerzonego.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}, \code{\link{procedura_1k_1w}},
\code{\link{przewidywanie_rasch}}, \code{\link{sprawdz_wyniki_skalowania}}
}

