% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skaluj_polichorycznie}
\alias{skaluj_polichorycznie}
\title{Skalowanie i łączenie kryteriów przy użyciu korelacji polichorycznych}
\usage{
skaluj_polichorycznie(dane, proceduraEG, korelacjaWiazki, nazwaSkalowania,
  ileKrokow = ifelse(is.null(korelacjaWiazki), 1000, nrow(korelacjaWiazki)),
  minMiara = NULL)
}
\arguments{
\item{dane}{ramka danych zawierająca wyniki egzaminów.}

\item{proceduraEG}{procedura skalowania.}

\item{korelacjaWiazki}{tablica zawierająca kolumny kr1 oraz kr2 z wartościami kryteriów do połącznia
oraz kolumnę numer określającą kolejność obliczeń korelacji. Jeżeli korelacjaWiazki=NULL
to funkcja liczy korelacje polichoryczne przed każdym połączeniem.}

\item{nazwaSkalowania}{ciąg znaków opisujący skalowanie.}

\item{ileKrokow}{liczba kolejnych skalowań, które ma wykonać funkcja.}

\item{minMiara}{wartość miary 1-R2, powyżej której obliczenia są przerywane.}
}
\value{
Funkcja zwraca obiekt klasy WynikSkalowania, który zawiera elementy:
\itemize{
\item \code{skalowania} - lista przeprowadzonych skalowań,
\item \code{korelacjaWiazki} - kryteria do połączenia,
\item \code{kolejnyIndeks} - indeks kolejnego skalowania,
\item \code{wartosciStartowe} - startowe wartości dla kolejnego skalowania,
\item \code{procedura} - procedura użyta przy ostatnim skalowaniu,
\item \code{nazwaSkalowania} - nazwa skalowania,
\item \code{dane} - dane użyte przy ostatnich obliczeniach,
\item \code{polaczenie} - wykonane połączenia oraz miary odpowiadające im,
}
}
\description{
Funkcja w pętli uruchamia skalowania. Każde kolejne skalowanie jest wykonywane
na zbiorze danych z dwoma połączonymi kryteriami. Łączenie kryteriów odbywa się na podstawie
korelacji polichorycznych.
}
\details{
Parametr ileKrokow określa liczbę połączeń, które zostaną wykonane na danych. Funkcja także wykonuje
skalowanie dla danych bez połączeń. Jeżeli ileKrokow wynosi 0 to wtedy zostaną wykonane obliczenia dla danych nie połączonych.
Definicja połączeń jest zawarta w tablicy korelacjaWiazki. Funkcja łączy kryteria określone przez początkowe
wiersze tej tablicy.
}

