% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skaluj_spr}
\alias{skaluj_spr}
\title{Procedury skalowania egzaminow.}
\usage{
skaluj_spr(daneWzorcowe, daneWszyscy, processors = 2)
}
\arguments{
\item{daneWzorcowe}{}

\item{daneWszyscy}{}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista z elementami:
\itemize{
  \item{\code{usunieteKryteria} wektor tekstowy z nazwami (pseudo)konstruktów, które
        zostały usunięte podczas skalowania wzorcowego.}
  \item{\code{parametry} data frame z wyestymowanymi parametrami modelu w jego
        ostatecznej postaci (tj. takiej, jak w ostatnim kroku skalowania wzorcowego
        i w jedynym kroku skalowania na wszystkich zdających).}
  \item{\code{oszacowania} data frame zawierający id_obserwacji i wyliczone
        oszacowania umiejętności dla wszystkich zdających.}
}
}
\description{
Procedura skalowania sprawdzianu.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}, \code{\link{procedura_spr}}
}

