% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skracanie_nazw_zmiennych.R
\name{skroc_nazwy_zmiennych}
\alias{skroc_nazwy_zmiennych}
\title{Skracanie nazw zmiennych.}
\usage{
skroc_nazwy_zmiennych(nazwy, dl)
}
\arguments{
\item{nazwy}{wektor tekstowy z nazwami do skrócenia}

\item{dl}{liczba całkowita maksymalna dopuszczalna długość nazwy}
}
\value{
wektor typu \code{character}
}
\description{
Funkcja stara się względnie inteligentnie skrócić podane nazwy do zadanej długości, zachowując ich unikalność.
}
\details{
Funkcja ma zaimplementowanych na sztywno kilka tłumaczeń-mapowań, następnie stara się usunąć "_" i "-", następnie usunąć samogłoski, ostatecznie nadaje nazwy "kolejne" bez związku z wcześniejszą nazwą.
}
