% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/skracanie_poziomy.R
\name{skroc_skale}
\alias{skroc_skale}
\title{Skracanie skal oceny}
\usage{
skroc_skale(x, maxLPozWyk = 5, minLiczebnPozWyk = 100,
  minOdsPozWyk = 0.05, print = TRUE, zrodloDanychODBC = "EWD")
}
\arguments{
\item{x}{data frame zawierający wyniki cześći egzaminu (typowo pobrane przy pomocy
funkcji \code{pobierz_czesc_egzaminu()} z pakietu \code{ZPD}) lub lista takich
data frame'ów}

\item{maxLPozWyk}{maksymalna liczba poziomów pytania}

\item{minLiczebnPozWyk}{minimalna liczebność obserwacji w każdym z poziomów}

\item{minOdsPozWyk}{minimalny odsetek obserwacji, które ma zawierać każdy poziom}

\item{print}{wartość logiczna - czy pokazywać informacje o skracaniu?}

\item{zrodloDanychODBC}{opcjonalnie nazwa źródła danych ODBC, dającego dostęp do
bazy (domyślnie "EWD")}
}
\value{
Data frame, pasujący swoją strukturą jako argument \code{elementy} do
funkcji \code{edytuj_skale} z pakietu \code{ZPD} lub lista takich data frame'ów.
}
\description{
Funkcja dokonuje skrótu skali oceny (pseudo)kryterium w oparciu o rozkład jego
wyników w grupie kalibracyjnej i informację o wszystkich możliwych wartościach,
jakie może przyjąć dane (pseudo)kryterium.
}

