% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sprawdz_wyniki_skalowania.R
\name{sprawdz_wyniki_skalowania}
\alias{sprawdz_wyniki_skalowania}
\title{Diagnostyka skalowania}
\usage{
sprawdz_wyniki_skalowania(model, nazwa = "")
}
\arguments{
\item{model}{lista zawierająca wyniki skalowania, typowo pojedynczy element listy
będącej wynikiem działania funkcji \code{\link{skaluj_egz_gimn}},
\code{\link{skaluj_egz_gimn_rasch}}, \code{\link{skaluj_matura_rasch}} lub wynik
działania funkcji \code{\link{skaluj_spr}}}

\item{nazwa}{ciag znaków - nazwa konstruktu}
}
\value{
funkcja nie zwraca żadnych wartości
}
\description{
Funkcja wyrysowuje zestaw wykresów diagnostycznych pozwalających przyjrzeć się wynikom
uzyskanym z modelu skalowania.
}
\seealso{
\code{\link{skaluj_egz_gimn}}, \code{\link{skaluj_egz_gimn_rasch}},
\code{\link{skaluj_matura_rasch}}, \code{\link{skaluj_spr}}
}

