% Generated by roxygen2 (4.0.2): do not edit by hand
\name{wstaw_do_skalowania_elementy}
\alias{wstaw_do_skalowania_elementy}
\title{Wstawianie danych do tablicy 'skalowania_elementy'}
\usage{
wstaw_do_skalowania_elementy(zrodloODBC, idSkali, kolejnosc, numerSkalowania,
  nazwaParametru, model, wartosc, odchylenie)
}
\arguments{
\item{zrodloODBC}{źródło danych ODBC.}

\item{idSkali}{numer opisujący id skali}

\item{kolejnosc}{liczba}

\item{numerSkalowania}{liczba}

\item{nazwaParametru}{ciąg znaków określający nazwę parametru}

\item{model}{ciąg znaków}

\item{wartosc}{liczba}

\item{odchylenie}{liczba opisująca odchylenie standardowe parametru.}
}
\value{
Funkcja nie zwraca żadnej wartości.
}
\description{
Funkcja wstawia jeden wiersz danych do tablicy 'skalowania_elementy'.
}

