% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zapisz_parametry_skalowania}
\alias{zapisz_parametry_skalowania}
\title{Zapisywanie parametrow skalowan.}
\usage{
zapisz_parametry_skalowania(nazwa_skali = NULL, id_testu = NULL, parametry,
  opis, estymacja, zrodloDanychODBC = "ewd_grzes")
}
\arguments{
\item{nazwa_skali}{nazwa skali. Kiedy nazwa skali przymuje wartość NULL nie jest brana pod
uwagę w wyszukiwaniu. Wartość domyślna to NULL.}

\item{id_testu}{id testu. Kiedy id testu przymuje wartość NULL nie jest brana pod
uwagę w wyszukiwaniu. Wartość domyślna to NULL.}

\item{parametry}{ramka danych o strukturze zgodnej z tą, w jakiej zwraca oszacowania parametrów funkcja skaluj().}

\item{opis}{opis estymacji}

\item{estymacja}{parametr opisujący metodę estymacji - musi być jedną z wartości występujących w tablicy sl_estymacje_parametrow.}

\item{zrodloDanychODBC}{string określający źródło danych. Wartość domyślna to 'EWD'.}
}
\value{
Funkcja nie zwraca żadnej wartości.
}
\description{
Funkcja zapisuje parametry skalowania do bazy danych
}

