% Generated by roxygen2 (4.0.1): do not edit by hand
\name{zapisz_skrocenia_do_bazy}
\alias{zapisz_skrocenia_do_bazy}
\title{Zapisywanie skróceń dla części egzaminu}
\usage{
zapisz_skrocenia_do_bazy(dane, idSkali, maxLPozWyk = 5,
  minLiczebnPozWyk = 100, minOdsPozWyk = 0.05, zrodloDanychODBC = "EWD")
}
\arguments{
\item{dane}{tablica z danymi z części egzaminu}

\item{idSkali}{numer skali}

\item{maxLPozWyk}{maksymalna liczba poziomów pytania}

\item{minLiczebnPozWyk}{minimalna liczebność obserwacji w każdym z poziomów}

\item{minOdsPozWyk}{minimalny odsetek obserwacji, które ma zawierać każdy poziom.}

\item{zrodloDanychODBC}{żródło danych}
}
\value{
funkcja nic nie zwraca
}
\description{
Funkcja zapisuje skrócenie do bazy.
}

