% Generated by roxygen2 (4.0.2): do not edit by hand
\name{zapisz_skrot_do_bazy}
\alias{zapisz_skrot_do_bazy}
\title{Zapisywanie skrócenia do bazy}
\usage{
zapisz_skrot_do_bazy(polaczenie, numerPytania, idSkali, czyPseudokryterium,
  zrodloDanychODBC = "EWD")
}
\arguments{
\item{polaczenie}{tablica poziomów (patrz opis wartości zwracanej przez funkcję \code{\link{okresl_skrocenie}} )}

\item{numerPytania}{numer kryterium lub pseudokryterium.}

\item{idSkali}{numer skali}

\item{czyPseudokryterium}{zmienna logiczna określająca czy mamy do czynienia z kryterium (FALSE) czy z pseudokryterium (TRUE)}

\item{zrodloDanychODBC}{źródło danych ODBC}
}
\value{
funkcja nic nie zwraca
}
\description{
Funkcja zapisuje skrócenie do bazy.
}

