% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skracanie_nazw_zmiennych.R
\name{zmien_nazwy_w_kroku_procedury}
\alias{zmien_nazwy_w_kroku_procedury}
\title{Skracanie nazw zmiennych.}
\usage{
zmien_nazwy_w_kroku_procedury(krok, nazwyDoZmiany)
}
\arguments{
\item{krok}{krok opisu procedury skalowania}

\item{nazwyDoZmiany}{lista, której elementami są zmienione (skrócone) nazwy zmiennych, a nazwami elementów listy są obecne nazwy zmiennych}
}
\value{
wektor
}
\description{
Funkcja przechodzi przez elementy \code{zmienne}, \code{wartosciStartowe}, \code{wartosciZakotwoczone}, \code{ograniczeniWartosci} kroku procedury i zamienia w nich nazwy zmiennych (typowo na skrócone).
}

