% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/skracanie_nazw_zmiennych.R
\name{czy_sie_nie_powtarza}
\alias{czy_sie_nie_powtarza}
\title{Obsluga ew. wystapienia duplikatow po skroceniu nazw zmiennych.}
\usage{
czy_sie_nie_powtarza(x, y)
}
\arguments{
\item{x}{wektor}

\item{y}{wektor}
}
\value{
wektor
}
\description{
Funkcja sprawdza, czy w pierwszym wektorze nie występują duplikaty. Jeśli występują, to podmienia je (zarówno pierwsze wystąpienie jak i powtórzenia) odpowiednimi wartościami z drugiego wektora.
}

