% Generated by roxygen2 (4.0.2): do not edit by hand
\name{korelacjePolihoryczne}
\alias{korelacjePolihoryczne}
\title{Korelacje polichoryczne dla tabeli danych}
\usage{
korelacjePolihoryczne(dane)
}
\arguments{
\item{dane}{dane w formie tabeli do obliczeń korelacji polichorycznych}
}
\value{
Funkcja zwraca wyniki w postaci macierzy kwadratowej o rozmiarze równym liczbie kolumn
tablicy danes.
}
\description{
Funkcja liczy macierz korelacji polichorycznych.
}

