% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skaluj_funkcje_pomocnicze.R
\name{obrob_out}
\alias{obrob_out}
\title{Parsowanie pliku wynikow Mplusa.}
\usage{
obrob_out(output, nazwyDoZmiany = NULL)
}
\arguments{
\item{output}{wektor tekstowy - wczytany plik poleceń Mplusa}

\item{nazwyDoZmiany}{lista, której elementami są zmienione (skrócone) nazwy zmiennych, a nazwami elementów listy są obecne nazwy zmiennych}
}
\value{
lista
}
\description{
Funkcja parsuje plik poleceń .out Mplusa i wyciąga z niego użyteczne informacje.
}
\details{
W elementach zwracanej listy znajdują się:
\itemize{
\item \code{podsumowanie} - ,
\item \code{dopasowanie} - ,
\item \code{parametry} - ,
\item \code{zapis} - .
\item \code{czas} - .
}
}
