% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedura_1k_1w.R
\name{procedura_1k_1w}
\alias{procedura_1k_1w}
\title{Procedury skalowania egzaminow.}
\usage{
procedura_1k_1w(
  nazwyZmiennych,
  nazwaKonstruktu = "f1",
  wartosciZakotwiczone = NULL,
  wartosciStartowe = NULL,
  rasch = FALSE,
  wieleGrup = NULL,
  maskaZmienne = "^([kp]_[[:digit:]]+|(s|t[[:digit:]]+)(nf|)_[[:alpha:]_]+)$",
  nigdyNieUsuwaj = NULL,
  processors = 3,
  integrPt = 20,
  usunWieleNaraz = FALSE,
  usunMimoKotwicy = FALSE,
  usunDyskrPonizej = 0.2
)
}
\arguments{
\item{nazwyZmiennych}{wektor tekstowy z nazwami zmiennych z data frame'a z danymi, na
których ma być prowadzona estymacja (funkcja sama wybierze tylko te, które pasują do
wyrażenia regularnego podanego w argumencie \code{maskaZmienne})}

\item{nazwaKonstruktu}{ciąg znaków - nazwa zmiennej opisującej mierzony konstrukt}

\item{wartosciZakotwiczone}{opcjonalnie data frame definiujący zakotwiczone wartości
parametrów modelu (p. \code{\link{skaluj}})}

\item{wartosciStartowe}{opcjonalnie data frame definiujący wartości startowe parametrów
modelu (p. \code{\link{skaluj}})}

\item{rasch}{opcjonalnie wartość logiczna - czy estymować model Rascha?}

\item{wieleGrup}{opcjonalnie wektor tekstowy z nazwami zmiennych (lub zmiennej)
definiujących podział na grupy w modelu wielogrupowym (w ramach grup uwolnione zostaną
wartości oczekiwane konstruktu )}

\item{maskaZmienne}{opcjonalnie wyrażenie regularne (ciąg znaków), które spełniają
nazwy tylko tych zmiennych, które mają być uwzględnione w skalowaniu}

\item{nigdyNieUsuwaj}{opcjonalnie wyrażenie regularne (ciąg znaków), które identyfikuje
te zmienne, które nie będą usuwane z modelu mimo spełnienia kryteriów usuwania}

\item{processors}{opcjonalnie liczba rdzeni do wykorzystania przy estymacji}

\item{integrPt}{opcjonalnie liczba punktów siatki do całkowania numerycznego (po
rozkładzie cechy ukrytej)}

\item{usunWieleNaraz}{opcjonalnie wartość logiczna - jeśli wiele zadań nie spełnia
kryterium \code{dyskryminacjaPonizej = 0.2}, to czy usuwać je wszystkie w jednym
kroku?}

\item{usunMimoKotwicy}{opcjonalnie wartość logiczna - jeśli podano wartości
zakotwiczone, to czy mimo to stosować kryteria usuwania zadań (przydaje się przy
maturze 2015-2016, gdy jednocześnie skalowana jest "stara" i "nowa" formuła)}

\item{usunDyskrPonizej}{opcjonalnie wartość liczbowa - próg wartości dyskryminacji
zadania, poniżej którego będą one usuwane z modelu skalowania}
}
\value{
lista, która zostanie użyta jako argument \code{opisProcedury} funkcji
\code{\link{skaluj}}
}
\description{
Procedura skalowania "jeden krok, jeden wymiar".
Funkcja przygotowuje opis bardzo prostej procedury skalowania (do użycia przez funkcję
\code{\link{skaluj}}), zawierającej tylko jeden krok, w ramach którego skalowany jest
jednowymiarowy konstrukt. Świetnie nadaje się do wykorzystania przy skalowaniu
sprawdzianu i egzaminu gimnazjalnego, które prowadzone są właśnie w taki prosty sposób.

Jeśli nie podano argumentu \code{wartosciZakotwiczone}, to zastosowane zostanie
kryterium usuwania zadań \code{dyskryminacjaPonizej = 0.2}. Jeśli został on podany, to
nie będą stosowane kryteria usuwania zadań oraz uwolnione zostaną wartość oczekiwana
i wariancja konstruktu.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}, \code{\link[EWDskale]{skaluj_spr}},
\code{\link[EWDskale]{skaluj_egz_gimn}}
}
