% Generated by roxygen2 (4.0.2): do not edit by hand
\name{przygotuj_parametry_IRT}
\alias{przygotuj_parametry_IRT}
\title{Wylicznie parametrów modeli IRT}
\usage{
przygotuj_parametry_IRT(skalowanie_laczenie)
}
\arguments{
\item{skalowanie_laczenie}{wyniki skalowania funkcją mirt.}
}
\value{
Funkcja zwraca listę zawierającą:
\itemize{
\item \code{theta} - umiejętności;
\item \code{a} - dyskryminacje;
\item \code{d} - średnia trudność;
}
}
\description{
Funkcja wylicza parametry modelu IRT.
}

