% Generated by roxygen2 (4.0.2): do not edit by hand
\name{skaluj_egz_gimn}
\alias{skaluj_egz_gimn}
\title{Procedury skalowania egzaminow.}
\usage{
skaluj_egz_gimn(daneWzorcowe, daneWszyscy, processors = 2)
}
\arguments{
\item{daneWzorcowe}{lista data frame'ów zawierających dane do skalowania wzorcowego}

\item{daneWszyscy}{lista data frame'ów zawierających zawierający dane wszystkich
zdających (do wyliczenia oszacowań umiejętności na podstawe parametrów ze skalowania
wzorcowego)}

\item{processors}{liczba rdzeni do wykorzystania przy estymacji}
}
\value{
lista z elementami:
\itemize{
  \item{\code{usunieteKryteria} lista wektorów tekstowych z nazwami (pseudo)kryteriów,
   które zostały usunięte podczas skalowania wzorcowego.}
  \item{\code{parametry} lista data frame'ów z wyestymowanymi parametrami modeli w ich
        ostatecznej postaci (tj. takiej, jak w ostatnim kroku skalowania wzorcowego
        i w jedynym kroku skalowania na wszystkich zdających).}
  \item{\code{oszacowania} lista data frame'ów zawierających id_obserwacji i wyliczone
        oszacowania umiejętności dla wszystkich zdających.}
}
}
\description{
Funkcja przeprowadza skalowanie wyników egzaminu gimnazjalnego.

Argumenty (poza ostatnim) muszą być listami, których elementy mają nazwy ze zbioru:
\code{"gh", "gh_h", "gh_p", "gm", "gm_p", "gm_m"} i są data frame'ami zawierającymi
dane z wynikami odpowiednich część egzaminu gimnazjalnego.

W przypadku nowej struktury egzaminu gimnazjalnego wystarczy podać wyniki testów,
dane do wyskalowania części jako całości zostaną połączone automatycznie, wewnątrz
funkcji.
}
\examples{
# chwilowo brak
}
\seealso{
\code{\link{skaluj}}, \code{\link{procedura_1k_1w}}
}

